// mobile/lib/shared/types.ts
export type PayoutMethod = 'bank_transfer' | 'cash_pickup' | 'mobile_wallet' | 'instapay';
export type SendMethod = 'bank_transfer' | 'debit_card' | 'credit_card' | 'apple_pay' | 'cash';
export type TransferSpeed = 'minutes' | 'hours' | 'same_day' | '1-2 days' | '2-3 days' | '3-5 days';